rename pointer events feedback
