+++ packages/@dcl/ecs/src/runtime/types.ts
@@ -12,25 +12,3 @@ export {
 } from '../components/legacy/Transform'
 export * from '../engine/component'
 export * from '../schemas/typing'
-
-// ColliderLayer is not used in the .pb definition because the field value is a number
-//   and it can have multiple values of this enum. It needs to be exposed here.
-export { BillboardMode } from '../components/generated/pb/decentraland/sdk/components/billboard.gen'
-export { AvatarModifierType } from '../components/generated/pb/decentraland/sdk/components/avatar_modifier_area.gen'
-export { CameraType } from '../components/generated/pb/decentraland/sdk/components/common/camera_type.gen'
-export {
-  Font,
-  TextAlignMode
-} from '../components/generated/pb/decentraland/sdk/components/common/texts.gen'
-export { RaycastQueryType } from '../components/generated/pb/decentraland/sdk/components/raycast.gen'
-
-export * from '../components/generated/pb/decentraland/sdk/components/ui_transform.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/avatar_attach.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/material.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/pointer_hover_feedback.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/pointer_events_result.gen'
-export * from '../components/generated/pb/decentraland/common/texture.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/raycast_result.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/animator.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/mesh_collider.gen'
-export * from '../components/generated/pb/decentraland/sdk/components/mesh_renderer.gen'
